bilet-backend/requirements.txt
annotated-doc
annotated-types
anyio
bcrypt
boto3
botocore
click
fastapi
h11
idna
iyzipay
jmespath
passlib
psycopg2-binary
pydantic
pydantic_core
PyJWT
python-dateutil
python-dotenv
python-multipart
s3transfer
six
SQLAlchemy
starlette
typing-inspection
typing_extensions
urllib3
uvicorn